refactor: give the clang AST cache-isolation helpers their own module
Fixes the real `ai-readiness` failure on this PR, which I had wrongly
assumed was the pre-existing ADR-049 one. Reading the job log instead:

  ERROR: [debt-no-growth] tests/test_clang_header_backend_integration.py:
  1678 lines exceeds adoption baseline 1604 and PR base 1600; move
  responsibility instead of raising the baseline

That file carries a `no_growth` baseline in `architecture/debt.yaml`, and my
earlier `_PruneSpy`/`_reset_ast_memo`/`_isolate_ast_cache` additions pushed it
78 lines past it. I had run `check_ai_readiness.py` but never
`check_architecture.py`, which AGENTS.md names as "the sharper gate for these
files and the one you will actually trip" -- so nothing local caught it.

Fixed the way AGENTS.md prescribes -- "move responsibility out to a properly-
owned module, never to trim the file to fit" -- not by raising the baseline.
`tests/_clang_ast_cache_isolation.py` now owns one responsibility: making two
differently-configured `dump()` calls genuinely independent, and proving the
second really ran. All three pieces belong together (a distinct on-disk cache
root, a cleared in-process memo -- a disk-cache miss alone does not force a
reparse -- and an assertion that the compared path executed), and they are
reusable by any other clang test needing the same guarantee.
`_METHOD_SHAPED_KINDS`/`_count_kinds` moved with them: proving what a
pruning-on run did to the tree is the same responsibility, the half a
differential test cannot infer from its own outputs.

The last few lines came from relocating the long "why" comments into that
module's docstring, where the responsibility now lives, leaving one-line
pointers at the call sites. 1678 -> 1601 lines, under the 1600 PR base and the
1604 adoption baseline; `check_architecture.py` reports 0 errors against both
`ARCHITECTURE_BASE=324815f7` and the CI-observed 851b1bd4.

Behaviour is unchanged and re-proven from the new module: all 26 tests in the
file pass, and reintroducing the shared cache root still fails both
differential tests on `spy.calls == 0` -- the extraction did not weaken the
assertion it exists for.
